scripts: missing dependency detection

+++ scripts/firewall.sh
@@ -1,12 +1,23 @@
 #!/bin/bash
-service=$(systemctl status firewall.service | grep 'Active:')
-echo $service
-status=$(echo "$service" | sed 's/     Active: active.*/on/ ; s/     Active: inactive.*/off/')
+
+service=$(systemctl status firewall.service 2>&1)
+
+# its important to note this script was made for nixos; other distros might not benefit from this script
+if [[ $(echo "$service" | sed 's/Unit firewall.service could not be found\./no/') == "no" ]]; then
+  echo "firewall not found"
+  exit 1
+fi
+
+active=$(echo "$service" | grep 'Active:')
+echo $active
+status=$(echo "$active" | sed 's/     Active: active.*/on/ ; s/     Active: inactive.*/off/')
 echo "$status"
+
 if [[ $status == "on" ]] ;then
   echo stopping
   systemctl stop firewall.service
   notify-send firewall off
+
 else
   if [[ $status == "off" ]] ;then
     echo turning on
